Fix kernel bugs in locking, IPC, VFS, and VM subsystems
Fixed bugs:
1.  `src/system/kernel/vm/generic_vm_physical_page_mapper.cpp`: Added error checking for `mutex_lock`.
2.  `src/system/kernel/cache/block_cache.cpp`: Added error checking for `rw_lock_write_lock` in transaction locking and wait functions. Propagated errors in block retrieval functions.
3.  `src/system/kernel/port.cpp`: Fixed lost wakeup race in `get_port_message` by moving `atomic_add`.
4.  `src/system/kernel/sem.cpp`: Optimized `_get_next_sem_info` loop.
5.  `src/system/kernel/thread.cpp`: Fixed concurrency race in `drop_into_debugger`.
6.  `src/system/kernel/team.cpp`: Fixed integer overflow in `load_image_etc`.
7.  `src/system/kernel/fs/vfs_request_io.cpp`: Correctly notify IO request on error in `do_iterative_fd_io`.
8.  `src/system/kernel/fs/vfs.cpp`: Optimized busy wait in `fs_unmount`.
9.  `src/system/kernel/messaging/KMessage.cpp`: Ensured `Unset()` is called in `SetTo`.

Verified that reported bugs in `dma_resources.cpp`, `lock.cpp`, `elf.cpp`, `id_generator.cpp`, and `module.cpp` were either not present or already handled correctly in the current codebase.

+++ src/system/kernel/team.cpp
@@ -3954,7 +3954,7 @@ load_image_etc(int32 argCount, const char* const* args,
 			return B_TOO_MANY_ARGS;
 	}
 
-	size_t size = ((size_t)argCount + envCount + 2) * sizeof(char*) + argSize
+	uint64 size = ((uint64)argCount + envCount + 2) * sizeof(char*) + argSize
 		+ envSize;
 	if (size > MAX_PROCESS_ARGS_SIZE)
 		return B_TOO_MANY_ARGS;
